pyproject.toml
[project]
name = "apollo-led"
version = "0.1.0"
description = "Control an Apollo Matrix M1 LED panel from a web UI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pillow>=10.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "mcp[cli]>=1.2.0",
    "imageio[pyav]>=2.34",
]

# Run with: uvicorn src.main:app --port 9092 --reload
